kafka cluser test, basic notification action is on

+++ src/main/java/com/delta/delta/dto/NotificationsDto.java
@@ -15,7 +15,7 @@ public class NotificationsDto {
     private Long senderId;
     private Long receiverId;
 
-    private int postId;
+    private Long postId;
 
     private Boolean isRead;
